sprawdzanie dostępności kodu produktu

+++ InventoryService/Service/ItemService.cs
@@ -34,7 +34,7 @@ namespace InventoryService.Service
         {
             _itemRepository.Delete(model);
         }
-        public Item GetItemByItemNumber (double data)
+        public Item GetItemByItemNumber (int data)
         {
             return _itemRepository.GetWhere(x => x.ItemNumber == data).FirstOrDefault();
         }
